getSystemPoi/requirements.txt
# Function dependencies, for example:
# package>=version
pymysql>=0.9
google.cloud.logging
